feat: add days_limit parameter to sale and expense forecast APIs and update state management, implement data on UI

+++ lib/data/model/response/forecast_response.dart
@@ -62,7 +62,7 @@ class MonthlyData {
 
 class SalesData {
   String? x;
-  int? y;
+  double? y;
 
   SalesData({this.x, this.y});
